requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
Automat==25.4.16
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
cssselect==1.3.0
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
executing==2.2.1
fastapi==0.119.1
fastjsonschema==2.21.2
filelock==3.20.0
graphviz==0.21
griffe==1.14.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
ipython==8.12.3
itemadapter==0.12.2
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mcp==1.17.0
mdurl==0.1.2
mistune==3.1.4
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.4
openai==1.109.1
openai-agents==0.3.3
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.5.0
postgrest==2.22.0
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.5.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyDispatcher==2.0.7
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyzmq==27.1.0
queuelib==1.8.0
realtime==2.22.0
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
rich==14.2.0
rpds-py==0.27.1
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
tinycss2==1.4.0
tldextract==5.3.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
w3lib==2.3.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
yarl==1.22.0
zope.interface==8.0.1
